classification/requirements.txt
matplotlib==3.4.2
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pytorch-lightning==1.3.3
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
torch==1.8.1
torchvision==0.9.1
tqdm==4.46.0
fastapi==0.63.0
uvicorn==0.13.4
requests==2.25.1
pydantic==1.8.1
joblib==1.0.1
starlette==0.13.8